Refactor: Moves findHighlightChunksInText to grafana/data
