+++ packages/grafana-data/src/types/index.ts
@@ -9,3 +9,4 @@ export * from './utils';
 export * from './valueMapping';
 export * from './displayValue';
 export * from './graph';
+export * from './text';
