+++ packages/grafana-data/src/types/text.ts
@@ -1,0 +1,6 @@
+export interface TextMatch {
+  text: string;
+  start: number;
+  length: number;
+  end: number;
+}
